# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  version: "1.0.0"
...
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authorization header using the Bearer
      scheme: bearer
      bearerFormat: JWT    
  schemas:
        Id:
          type: string
          description: The unique identifier for the user.
          readOnly: True
          pattern: "{09az}{8}-{09az}{4}-{09az}{4}-{09az}{4}-{09az}{12}"
          minLength: 36
          maxLength: 36
        Username:
          type: string
          description: The username of the user.
          example: "mario"
          readOnly: True
          pattern: '^.*$'
          minLength: 3
          maxLength: 20
        Title: 
          type: string
          description: The title of the photo.
          pattern: '^.*?$'
          minLength: 0
          maxLength: 30
        File:
          description: The photo file to upload.
          type: string
          pattern: '^[01]*$'
          format: binary
          minLength: 0
          maxLength: 10000000
        Date:
          description: A date represented by year, month, and day components(YYYY-MM-DD).
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          minLength: 10
          maxLength: 10
...
        Path:
          description: The paths of the photo.
          type: string
          pattern: '^.*?$'
          minLength: 0
          maxLength: 300
...
        Photo:
          title: Photo
          description: Photo posted by a user
          type: object
          properties:
            photoId:
              $ref: "#/components/schemas/Id"
            title:
              $ref: "#/components/schemas/Title"
            photoPath:
              $ref: "#/components/schemas/Path"
            date:
              $ref: "#/components/schemas/Date"
            author:
              $ref: "#/components/schemas/Id"
          required:
            - photoId
            - title
            - photoPath
            - date
            - author
...
security:
  - bearerAuth: []
paths:
...
  /photos/: #uploadPhoto
    post:
      tags: ["photo"]
      summary: upload a photo
      description: |
        Upload a photo.
        The user must be already logged in.
      operationId: uploadPhoto
      parameters: 
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" } 
      requestBody:
        description: Photo and description
        required: true
        content:
          multipart/form-data:
            schema:
              type: array 
              minItems: 3 
              maxItems: 3
              description: Photo, description and author
              items:
                type: object
                properties:
                  title:
                    $ref: "#/components/schemas/Title"
                  photo:
                    $ref: "#/components/schemas/File"
                  userId:
                    $ref: "#/components/schemas/Id"
                  
        
      responses:
        "201":
          description: Photo succesfully uploaded
          content:
            multipart/form-data:
              schema: { $ref: "#/components/schemas/Photo" }
        "400":
          description: |
            The photo is not in the correct format
          
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /follow/{username}: #unfollowUser
    delete:
      tags: ["following"]
      summary: Unfollow a user
      description: Unfollow a user
      operationId: unfollowUser
      parameters:
        - name: userId
          in: query
          description: The identifier of the user.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: username
          in: path
          description: The username of the user to unfollow.
          required: true
          schema: { $ref: "#/components/schemas/Username"}
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      responses:
        "204":
          description: User succesfully unfollowed
        "403":
          description: |
            User not authorized to unfollow himself 
            or a user not followed
            or user blocked by the user tried to unfollow
        "404":
          description: User to unfollow not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /ban/{username}: #unbanUser
    delete:
      tags: ["ban"]
      operationId: unbanUser
      summary: Unban a user
      description: |
        Unban a user.
        The user must be already logged in.
      parameters:
        - name: userId
          in: query
          description: The identifier of the user who want to unban another user.
          required: true
          schema: { $ref: "#/components/schemas/Id"}
        - name: username
          in: path
          description: The username of the user unbanned.
          required: true
          schema: { $ref: "#/components/schemas/Username"}
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" }
      responses:
        "204":
          description: User succesfully unfollowed
        "403":
          description: |
            User not authorized to unban himself 
            or user already unbanned
            or user blocked by the user tried to unban
        "404":
          description: User to unfollow not found, maybe deleted the account
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /likes/{photoId}: #unlikePhoto
    delete: 
      tags: ["like"]
      summary: Unlike a photo.
      description: |
        Unlike a photo.
        The user must be already logged in.
      operationId: unlikePhoto
      parameters:
        - name: photoId
          in: path
          description: The identifier of the photoId to unlike
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" } 
        - name: userId
          in: query
          description: The identifier of the user used as a form of authentication.
          required: true
          schema: { $ref: "#/components/schemas/Id"}
      responses:
        "204":
          description: Photo succesfully unliked
        "403":
          description: |
            photo deleted, blocked by the owner of the photo (photo retrieved by a link or by other form of sharing)
        "404":
          description: Photo not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /comments/{commentId}: #uncommentPhoto
    delete:
      tags: ["comment"]
      summary: Delete a comment from a photo
      description: |
        Delete a comment from a photo.
        The user must be already logged in.
      operationId: uncommentPhoto
      parameters:
        - name: commentId
          in: path
          description: The identifier of the comment
          required: true
          schema: { $ref: "#/components/schemas/Id" }
        - name: token
          in: query
          description: The token for the authentication method.
          required: true
          schema: { $ref: "#/components/schemas/Id" } 
        - name: userId
          in: query
          description: The identifier of the user used as a form of authentication.
          required: true
          schema: { $ref: "#/components/schemas/Id"}
      responses:
        "204":
          description: Comment succesfully deleted"
        "403":
          description: |
            User not authorized to delete a comment, not the owner of the photo
            or cannot access to the photo since the user is blocked by the owner of the photo (photo retrieved by a link or by other form of sharing)
        "404":
          description: |
            Comment not found or
            Photo not found, maybe deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"